Updates to file size handler

+++ src/commonStructures.h
@@ -18,7 +18,7 @@ struct filterStruct{
 
 struct dataPackage {
 	char* dataPointer = nullptr;
-	int size = -1;
+	long size = -1;
 };
